windows/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
MetaTrader5==5.0.45
pandas==2.1.0
PyYAML==6.0.1
